feat: mapeado os usuarios do sistema para o orm
